requirements.txt
streamlit
requests
PyPDF2
python-dotenv
fpdf
python-docx